+++ backend/src/main/java/com/team3/airdnd/accommodation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package com.team3.airdnd.accomodation.domain;
+package com.team3.airdnd.accommodation.domain;
 
 import com.team3.airdnd.user.domain.User;
 import jakarta.persistence.*;
