+++ backend/src/main/java/com/team3/airdnd/accommodation/repository/AccommodationRepository.java
@@ -1,6 +1,6 @@
-package com.team3.airdnd.accomodation.repository;
+package com.team3.airdnd.accommodation.repository;
 
-import com.team3.airdnd.accomodation.domain.Accommodation;
+import com.team3.airdnd.accommodation.domain.Accommodation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
