+++ backend/src/main/java/com/team3/airdnd/accommodation/service/AccomodationService.java
@@ -1,4 +1,4 @@
-package com.team3.airdnd.accomodation.service;
+package com.team3.airdnd.accommodation.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
