--- backend/src/main/java/com/team3/airdnd/accomodation/dto/AccommodationRequestDto.java
@@ -1,4 +1,0 @@
-package com.team3.airdnd.accomodation.dto;
-
-public class AccommodationRequestDto {
-}
