--- backend/src/main/java/com/team3/airdnd/accomodation/dto/AccommodationResponseDto.java
@@ -1,4 +1,0 @@
-package com.team3.airdnd.accomodation.dto;
-
-public class AccommodationResponseDto {
-}
