+++ backend/src/main/java/com/team3/airdnd/payment/domain/Payment.java
@@ -1,6 +1,6 @@
 package com.team3.airdnd.payment.domain;
 
-import com.team3.airdnd.accomodation.domain.Reservation;
+import com.team3.airdnd.accommodation.domain.Reservation;
 import jakarta.persistence.*;
 import lombok.*;
